requirements.txt
asgiref==3.4.0
autopep8==1.5.7
Django==3.2.4
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2

# django rest framework jwt token
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
django-allauth==0.51.0
dj-rest-auth==2.2.5
django-rest-auth==0.9.5
markdown==3.4.1


# S3 Bucket
boto3
django-storages
pillow
